Restructure of project
