+++ examples/simple.gu
@@ -19,5 +19,6 @@
 
 let i: float = 2.0;
 let j: int = 1;
+let k: bool = true;
 
-(i + i) / 2 || false;
+((i - i) * 2) >= j || k && false;
