+++ src/compiler/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod lexer;
+pub(crate) mod parser;
+pub(crate) mod semantics;
+pub(crate) mod translator;
