+++ src/compiler/parser/mod.rs
@@ -1,224 +1,13 @@
-use std::fmt::Display;
+pub mod ast;
 
 use chumsky::{
     prelude::{filter_map, just, skip_then_retry_until, recursive, Simple, nested_delimiters, choice},
     Parser, select
 };
 
-use crate::lexer::{Token, Span, Op};
+use super::lexer::{Token, Span, Op};
 
-pub(crate) type Spanned<T> = (T, Span);
-pub(crate) type Block = Vec<Spanned<Stmt>>;
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum Stmt {
-    Decl(Decl),
-    Assign {
-        to: Field, 
-        value: Spanned<Expr>,
-    },
-    Cond {
-        if_: Spanned<Expr>,
-        then: Block,
-        elif: Vec<(Spanned<Expr>, Block)>,
-        else_: Option<Block>,
-    },
-    Loop(Loop),
-    Expr(Spanned<Expr>),
-    Return(Spanned<Expr>),
-    Error,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) struct Field {
-    name: String,
-    child: Option<Box<Self>>,
-}
-impl ToString for Field {
-    fn to_string(&self) -> String {
-        let mut s = String::new();
-        s.push_str(&self.name);
-        if let Some(ref child) = self.child {
-            s.push_str(".");
-            s.push_str(&child.to_string());
-        }
-        s
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum Decl {
-    Let { 
-        name: String,
-        type_: Type,
-        value: Option<Spanned<Expr>>,
-    },
-    Fun(Fun),
-    Class {
-        name: String,
-        inherits: Option<String>,
-        implements: Option<String>,
-        has: Vec<Spanned<Var>>,
-        does: Vec<Spanned<Fun>>,
-    },
-    Interface {
-        name: String,
-        should_do: Vec<Spanned<FunSignature>>,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) struct Var {
-    pub name: String,
-    pub type_: Type,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) struct Fun {
-    pub name: String,
-    pub params: Vec<Spanned<Var>>,
-    pub ret_type: Option<Type>,
-    pub body: Block,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) struct FunSignature {
-    pub name: String,
-    pub params: Vec<Spanned<Var>>,
-    pub ret_type: Option<Type>,
-}
-
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum Loop {
-    While {
-        cond: Spanned<Expr>,
-        body: Block,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum Expr {
-    Error,
-    Call {
-        fun: Box<Spanned<Self>>, 
-        args: Vec<Spanned<Self>>,
-    },
-    Binary {
-        lhs: Box<Spanned<Self>>, 
-        op: BinOp, 
-        rhs: Box<Spanned<Self>>,
-    },
-    Unary { 
-        op: UnOp,
-        rhs: Box<Spanned<Self>>,
-    },
-    Constant(Literal),
-    Ident(String),
-    Array(Vec<Spanned<Self>>),
-    Parenthesized(Box<Spanned<Self>>),
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum BinOp {
-    //// Structural
-    Chain,
-
-    //// Arithmetic
-    Add,
-    Sub,
-    Mul,
-    Div,
-
-    //// Comparison
-    Eq,
-    Ne,
-    Lt,
-    Gt,
-    Lte,
-    Gte,
-
-    //// Boolean
-    And,
-    Or,
-}
-impl Display for BinOp {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            BinOp::Chain => write!(f, "."),
-            BinOp::Add => write!(f, "ADD"),
-            BinOp::Sub => write!(f, "SUB"),
-            BinOp::Mul => write!(f, "MUL"),
-            BinOp::Div => write!(f, "DIV"),
-            BinOp::Eq => write!(f, "EQ"),
-            BinOp::Ne => write!(f, "NEQ"),
-            BinOp::Lt => write!(f, "LT"),
-            BinOp::Gt => write!(f, "GT"),
-            BinOp::Lte => write!(f, "LTE"),
-            BinOp::Gte => write!(f, "GTE"),
-            BinOp::And => write!(f, "AND"),
-            BinOp::Or => write!(f, "OR"),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum UnOp {
-    Not,
-    Neg,
-}
-impl Display for UnOp {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            UnOp::Not => write!(f, "NOT"),
-            UnOp::Neg => write!(f, "NEG"),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum Literal {
-    Int(i64),
-    Float(f64),
-    Bool(bool),
-    Char(char),
-    String(String),
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum Type {
-    //// Primitive
-    Int,
-    Float,
-    Bool,
-    Char,
-
-    //// Compound
-    Array(Box<Self>),
-    String,
-
-    //// Custom
-    Custom(String),
-
-    //// Error handling
-    Error,
-}
-
-impl Display for Type {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Type::Int => write!(f, "int"),
-            Type::Float => write!(f, "float"),
-            Type::Bool => write!(f, "bool"),
-            Type::Char => write!(f, "char"),
-            Type::Array(t) => write!(f, "[{:?}]", t),
-            Type::String => write!(f, "string"),
-            Type::Custom(name) => write!(f, "{}", name),
-            Type::Error => todo!(),
-        }
-    }
-}
-
+use self::ast::*;
 
 pub(crate) fn parser() -> impl Parser<Token, Vec<Spanned<Stmt>>, Error = Simple<Token>> {
     let ident = select! { Token::Ident(name) => name.clone() }.labelled("identifier");
@@ -624,7 +413,7 @@ mod tests {
 
     use chumsky::Stream;
 
-    use crate::lexer::lexer;
+    use crate::compiler::lexer::lexer;
 
     use super::*;
