+++ src/compiler/semantics/mod.rs
@@ -1,12 +1,16 @@
 
-use std::fmt::{Debug, Display};
+pub mod item;
+pub mod semantic_cube;
 
-use crate::{
-  directory::{self, Scope, Kind, Item}, 
-  parser::{Stmt, Decl, Expr, Loop, Type, Literal, BinOp, UnOp, Fun, Block}, 
-  lexer::Span,
-  semantic_cube::resolve,
-};
+use std::{fmt::{Debug, Display}, ops::Range};
+
+use crate::directory::{self, Dir};
+
+use self::item::{Kind, Item};
+
+use super::{parser::ast::*, semantics::semantic_cube::resolve, lexer::Span};
+
+pub(crate) type Scope = Dir<Item>;
 
 // static mut SCOPE: directory::Scope = directory::Scope::new();
 #[derive(Debug, Clone, PartialEq)]
@@ -38,7 +42,8 @@ impl Display for Error {
   }
 }
 
-type Result<T> = std::result::Result<T, Error>;
+pub(self) type Result<T> = std::result::Result<T, Spanned<Error>>;
+
 fn map_dir_errs(err: directory::Error) -> Error {
   match err {
     directory::Error::MismatchingTypes(id, l_type, r_type) => Error::TypeMismatch(id, l_type, r_type),
@@ -57,7 +62,7 @@ pub(crate) fn eval_expr(expr: Expr, span: Span, scope: &Scope) -> Result<Type> {
       let lhs_type = eval_expr(lhs.0, lhs.1, scope)?;
       let rhs_type = eval_expr(rhs.0, rhs.1, scope)?;
       
-      resolve(&lhs_type, &op, &rhs_type).map_err(|_| Error::Incompatible(op, lhs_type, rhs_type))
+      resolve(&lhs_type, &op, &rhs_type).map_err(|_| (Error::Incompatible(op, lhs_type, rhs_type), span))
     },
     Expr::Unary { op, rhs } => todo!(),
     Expr::Constant(literal) => match literal {
@@ -68,28 +73,28 @@ pub(crate) fn eval_expr(expr: Expr, span: Span, scope: &Scope) -> Result<Type> {
         Literal::String(_) => Ok(Type::String),
     },
     Expr::Ident(id) => {
-      let item = scope.get(id.as_str()).map_err(map_dir_errs)?;
+      let item = scope.get(id.as_str()).map_err(|e| (map_dir_errs(e), span.clone()))?;
       if item.kind != Kind::Var {
-        return Err(Error::NotAVariable(id.to_string(), item.kind.clone()));
+        return Err((Error::NotAVariable(id.to_string(), item.kind.clone()), span.clone()));
       }
-      item.type_.clone().ok_or_else(|| Error::UntypedVariable(id.to_string()))
+      item.type_.clone().ok_or_else(|| (Error::UntypedVariable(id.to_string()), span.clone()))
     },
     Expr::Array(exprs) => {
       
       // Empty array declarations are not allowed
       if exprs.is_empty() {
-        return Err(Error::EmptyArray);
+        return Err((Error::EmptyArray, span.clone()));
       }
 
       // Evaluate items in the array
       let mut types = Vec::new();
-      for expr in exprs {
-        types.push(eval_expr(expr.0, expr.1, scope)?);
+      for (expr, span) in exprs {
+        types.push(eval_expr(expr, span.clone(), scope)?);
       }
 
       // All must be of the same type
       if types.iter().any(|t| *t != types[0]) {
-          return Err(Error::HeterogenousArray);
+          return Err((Error::HeterogenousArray, span.clone()));
       }
 
       Ok(Type::Array(Box::new(types[0].clone())))
@@ -110,10 +115,10 @@ pub(crate) fn eval_stmt(stmt: Stmt, span: Span, scope: &mut Scope) -> Result<()>
           if let Some(value) = value {
             let value_type = eval_expr(value.0, value.1, &scope)?;
             if type_ != value_type {
-              return Err(Error::TypeMismatch(name.to_string(), type_, value_type));
+              return Err((Error::TypeMismatch(name.to_string(), type_, value_type), span.clone()));
             }
           }
-          scope.create(Item::new(name, Kind::Var, type_)).map_err(map_dir_errs)
+          scope.create(Item::new(name, Kind::Var, type_)).map_err(|e| (map_dir_errs(e), span.clone()))
         },
 
         Decl::Fun(Fun { name, params, ret_type, body }) => {
@@ -123,16 +128,16 @@ pub(crate) fn eval_stmt(stmt: Stmt, span: Span, scope: &mut Scope) -> Result<()>
           for param in params {
             subscope
               .create(Item::new(param.0.name, Kind::Var, param.0.type_))
-            .map_err(map_dir_errs)?;
+            .map_err(|e| (map_dir_errs(e), span.clone()))?;
           }
 
           for (stmt, span) in body {
-            eval_stmt(stmt, span, &mut subscope)?;
+            eval_stmt(stmt, span.clone(), &mut subscope)?;
           }
 
           *scope = subscope.drop();
 
-          scope.create(Item { id: name, kind: Kind::Fun, type_: ret_type }).map_err(map_dir_errs)
+          scope.create(Item { id: name, kind: Kind::Fun, type_: ret_type }).map_err(|e| (map_dir_errs(e), span.clone()))
         },
 
         Decl::Class { name, inherits, implements, has, does } => todo!(),
@@ -145,16 +150,16 @@ pub(crate) fn eval_stmt(stmt: Stmt, span: Span, scope: &mut Scope) -> Result<()>
       let value_type = eval_expr(value.0, value.1, &scope)?;
       let to_type = scope
         .get(to.to_string().as_str())
-        .map_err(map_dir_errs)?
+        .map_err(|e| (map_dir_errs(e), span.clone()))?
         .type_
         .clone()
         .expect("Expressions should have a type");
 
       if value_type != to_type {
-        return Err(Error::TypeMismatch(to.to_string(), to_type, value_type));
+        return Err((Error::TypeMismatch(to.to_string(), to_type, value_type), span.clone()));
       }
 
-      scope.update(Item::new(to.to_string(), Kind::Var, value_type)).map_err(map_dir_errs)
+      scope.update(Item::new(to.to_string(), Kind::Var, value_type)).map_err(|e| (map_dir_errs(e), span.clone()))
     },
 
     Stmt::Cond { if_, then, elif, else_ } => todo!(),
@@ -175,10 +180,10 @@ pub(crate) fn eval_stmt(stmt: Stmt, span: Span, scope: &mut Scope) -> Result<()>
 }
 
 
-pub(crate) fn semantic_analysis(stmts: Block) -> std::result::Result<(), Vec<Error>> {
+pub(crate) fn semantic_analysis(stmts: Block) -> std::result::Result<(), Vec<Spanned<Error>>> {
 
-  let mut scope = directory::Scope::new();
-  let errors: Vec<Error> = stmts
+  let mut scope = Scope::new();
+  let errors: Vec<Spanned<Error>> = stmts
     .iter()
     .map(|(stmt, span)| eval_stmt(stmt.clone(), span.clone(), &mut scope))
     .filter_map(|res| res.err())
@@ -194,13 +199,11 @@ pub(crate) fn semantic_analysis(stmts: Block) -> std::result::Result<(), Vec<Err
 #[cfg(test)]
 mod tests {
 
-  use crate::parser::Var;
-
 use super::*;
 
   #[test]
   fn test_eval_expr() {
-    let scope = directory::Scope::new();
+    let scope = Scope::new();
     let expr = Expr::Binary {
       lhs: Box::new((Expr::Constant(Literal::Int(1)), 0..0)),
       op: BinOp::Add,
@@ -212,7 +215,7 @@ use super::*;
 
   #[test]
   fn test_eval_expr_with_scope() {
-    let mut scope = directory::Scope::new();
+    let mut scope = Scope::new();
     scope.create(Item::new("x".to_string(), Kind::Var, Type::Int)).unwrap();
     let expr = Expr::Ident("x".to_string());
     let result = eval_expr(expr, 0..0, &scope);
@@ -221,7 +224,7 @@ use super::*;
 
   #[test]
   fn test_decl_var() {
-    let mut scope = directory::Scope::new();
+    let mut scope = Scope::new();
     let stmt = Stmt::Decl(Decl::Let {
       name: "x".to_string(),
       type_: Type::Int,
@@ -231,7 +234,7 @@ use super::*;
     let result = eval_stmt(stmt, 0..0, &mut scope);
     assert_eq!(result, Ok(()));
 
-    let mut scope = directory::Scope::new();
+    let mut scope = Scope::new();
     let stmt = Stmt::Decl(Decl::Let {
       name: "x".to_string(),
       type_: Type::String,
@@ -244,7 +247,7 @@ use super::*;
 
   #[test]
   fn test_decl_fun() {
-    let mut scope = directory::Scope::new();
+    let mut scope = Scope::new();
     let stmt = Stmt::Decl(Decl::Fun(Fun{
       name: "f".to_string(),
       params: vec![(Var {name: "x".to_string(), type_: Type::Int}, 0..0)],
